requirements.txt
anyio==3.6.2
bcrypt==4.0.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
click==8.1.3
config==0.5.1
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.92.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
orjson==3.8.6
pydantic==1.10.5
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
pymongo==4.3.3
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.20.0
uvloop==0.17.0
validate_email==1.3
watchfiles==0.18.1
websockets==10.4
